requirements.txt
fastapi
uvicorn
tensorflow
pillow
requests
python-dotenv
